requirements.txt
numpy
Keras 
Pillow
tensorflow
streamlit==1.1.0
